Fix absent qemu-img commit -b, #PSBM-39075
This option is only present in qemu-img 2.2+.

It means that we can specify an image in backing chain
to which all changes will be committed.
If this option is absent then we have to commit each image
previous until we reach the specified base.
This approach is slower and consumes more disk space.

Added some error handling on json parsing.